generated_projects/local_discord_bot/local_discord_bot/requirements.txt
discord.py==2.0.0
aiohttp==3.8.1